Rename describe.rb to describe_base.rb
