+++ lib/inspec/describe_base.rb
